fix: fixed bug with update function, it updated based on wallet not userid

+++ src/Explorer.Payments.API/Dtos/ShoppingCart/WalletDTO.cs
@@ -5,6 +5,6 @@ namespace Explorer.Payments.API.Dtos.ShoppingCart
     {
         public long Id { get; set; }
         public long TouristId { get; set; }
-        public decimal Balance { get; set; }
+        public double Balance { get; set; }
     }
 }
